+++ src/Components/Broker/YFinanceInterface.py
@@ -1,0 +1,3 @@
+class YFinanceInterface:
+    def __init__(self):
+        super().__init__()