+++ test/test_ig_interface.py
@@ -8,7 +8,14 @@ currentdir = os.path.dirname(os.path.abspath(inspect.getfile(inspect.currentfram
 parentdir = os.path.dirname(currentdir)
 sys.path.insert(0, "{}/src".format(parentdir))
 
-from Components.IGInterface import IGInterface
+from Components.Configuration import Configuration
+from Components.Broker.BrokerFactory import InterfaceNames
+from Components.Broker.IGInterface import IGInterface
+from Components.Utils import TradeDirection, Interval
+from Interfaces.Position import Position
+from Interfaces.Market import Market
+from Interfaces.MarketMACD import MarketMACD
+from Interfaces.MarketHistory import MarketHistory
 from common.MockRequests import (
     ig_request_login,
     ig_request_set_account,
@@ -26,87 +33,52 @@ from common.MockRequests import (
 
 @pytest.fixture
 def config():
-    """
-    Returns a dict with config parameter for ig_interface
-    """
-    return {
-        "ig_interface": {
-            "order_type": "MARKET",
-            "order_size": 1,
-            "order_expiry": "DFB",
-            "order_currency": "GBP",
-            "order_force_open": True,
-            "use_g_stop": True,
-            "use_demo_account": True,
-            "controlled_risk": True,
-            "paper_trading": False,
-        }
-    }
+    with open("test/test_data/config.json", "r") as f:
+        config = json.load(f)
+        # Inject ig_interface as active interface in the config file
+        config["stocks_interface"]["active"] = InterfaceNames.IG_INDEX.value
+        config["account_interface"]["active"] = InterfaceNames.IG_INDEX.value
+        return Configuration(config)
 
 
 @pytest.fixture
-def credentials():
-    """
-    Returns a dict with credentials parameters
-    """
-    return {
-        "username": "user",
-        "password": "pwd",
-        "api_key": "12345",
-        "account_id": "12345",
-        "av_api_key": "12345",
-    }
-
-
-@pytest.fixture
-def ig(requests_mock, config, credentials):
+def ig(requests_mock, config):
     """
     Returns a instance of IGInterface
     """
     ig_request_login(requests_mock)
     ig_request_set_account(requests_mock)
-    return IGInterface(config, credentials)
+    return IGInterface(config)
 
 
-def test_init(ig, config):
-    assert ig.orderType == config["ig_interface"]["order_type"]
-    assert ig.orderSize == config["ig_interface"]["order_size"]
-    assert ig.orderExpiry == config["ig_interface"]["order_expiry"]
-    assert ig.orderCurrency == config["ig_interface"]["order_currency"]
-    assert ig.orderForceOpen == config["ig_interface"]["order_force_open"]
-    assert ig.useGStop == config["ig_interface"]["use_g_stop"]
-    assert ig.useDemo == config["ig_interface"]["use_demo_account"]
-    assert ig.paperTrading == config["ig_interface"]["paper_trading"]
-
-
-# No need to use requests_mock fixture as "ig" already does that
-def test_authenticate(ig, credentials):
+# No need to use requests_mock as "ig" fixture already does that
+def test_authenticate(ig):
     # Call function to test
-    result = ig.authenticate(credentials)
+    result = ig.authenticate()
     # Assert results
     assert ig.authenticated_headers["CST"] == "mock"
     assert ig.authenticated_headers["X-SECURITY-TOKEN"] == "mock"
     assert result == True
 
 
-def test_authenticate_fail(requests_mock, ig, credentials):
+def test_authenticate_fail(requests_mock, ig):
     ig_request_login(requests_mock, fail=True)
     ig_request_set_account(requests_mock, fail=True)
     # Call function to test
-    result = ig.authenticate(credentials)
+    result = ig.authenticate()
     # Assert results
     assert result == False
 
 
 # No need to use requests_mock fixture
-def test_set_default_account(ig, credentials):
-    result = ig.set_default_account(credentials["account_id"])
+def test_set_default_account(ig):
+    result = ig.set_default_account("mock")
     assert result == True
 
 
-def test_set_default_account_fail(requests_mock, ig, credentials):
+def test_set_default_account_fail(requests_mock, ig):
     ig_request_set_account(requests_mock, fail=True)
-    result = ig.set_default_account(credentials["account_id"])
+    result = ig.set_default_account("mock")
     assert result == False
 
 
@@ -116,14 +88,13 @@ def test_get_account_balances(requests_mock, ig):
     assert balance is not None
     assert deposit is not None
     assert balance == 16093.12
-    assert deposit == 0.0
+    assert deposit == 10000.0
 
 
 def test_get_account_balances_fail(requests_mock, ig):
     ig_request_account_details(requests_mock, fail=True)
-    balance, deposit = ig.get_account_balances()
-    assert balance is None
-    assert deposit is None
+    with pytest.raises(RuntimeError) as e:
+        balance, deposit = ig.get_account_balances()
 
 
 def test_get_open_positions(ig, requests_mock):
@@ -132,14 +103,15 @@ def test_get_open_positions(ig, requests_mock):
     positions = ig.get_open_positions()
 
     assert positions is not None
-    assert "positions" in positions
+    assert isinstance(positions, list)
+    assert len(positions) > 0
+    assert isinstance(positions[0], Position)
 
 
 def test_get_open_positions_fail(ig, requests_mock):
     ig_request_open_positions(requests_mock, fail=True)
-    positions = ig.get_open_positions()
-
-    assert positions is None
+    with pytest.raises(RuntimeError) as e:
+        positions = ig.get_open_positions()
 
 
 def test_get_market_info(ig, requests_mock):
@@ -147,59 +119,59 @@ def test_get_market_info(ig, requests_mock):
     info = ig.get_market_info("mock")
 
     assert info is not None
-    assert "instrument" in info
-    assert "snapshot" in info
-    assert "dealingRules" in info
+    assert isinstance(info, Market)
 
 
 def test_get_market_info_fail(ig, requests_mock):
     ig_request_market_info(requests_mock, fail=True)
-    info = ig.get_market_info("mock")
-    assert info is None
+    with pytest.raises(RuntimeError) as e:
+        info = ig.get_market_info("mock")
 
 
 def test_search_market(ig, requests_mock):
+    ig_request_market_info(requests_mock)
     ig_request_search_market(requests_mock)
     markets = ig.search_market("mock")
 
     assert markets is not None
+    assert isinstance(markets, list)
     assert len(markets) == 8
-    assert "epic" in markets[0]
-    assert "expiry" in markets[0]
-    assert "bid" in markets[0]
-    assert "offer" in markets[0]
+    assert isinstance(markets[0], Market)
 
 
 def test_search_market_fail(ig, requests_mock):
     ig_request_search_market(requests_mock, fail=True)
-    markets = ig.search_market("mock")
-    assert markets is None
+    with pytest.raises(RuntimeError) as e:
+        markets = ig.search_market("mock")
 
 
 def test_get_prices(ig, requests_mock):
+    ig_request_market_info(requests_mock)
     ig_request_prices(requests_mock)
-    p = ig.get_prices("mock", "mock", "mock")
+    p = ig.get_prices(ig.get_market_info("mock"), Interval.HOUR, 10)
     assert p is not None
-    assert "prices" in p
+    assert isinstance(p, MarketHistory)
+    assert len(p.dataframe) > 0
 
 
 def test_get_prices_fail(ig, requests_mock):
+    ig_request_market_info(requests_mock)
     ig_request_prices(requests_mock, fail=True)
-    p = ig.get_prices("mock", "mock", "mock")
-    assert p is None
+    with pytest.raises(RuntimeError) as e:
+        p = ig.get_prices(ig.get_market_info("mock"), Interval.HOUR, 10)
 
 
 def test_trade(ig, requests_mock):
     ig_request_trade(requests_mock)
     ig_request_confirm_trade(requests_mock)
-    result = ig.trade("mock", "BUY", 0, 0)
+    result = ig.trade("mock", TradeDirection.BUY, 0, 0)
     assert result
 
 
 def test_trade_fail(ig, requests_mock):
     ig_request_trade(requests_mock, fail=True)
     ig_request_confirm_trade(requests_mock, fail=True)
-    result = ig.trade("mock", "BUY", 0, 0)
+    result = ig.trade("mock", TradeDirection.BUY, 0, 0)
     assert result == False
 
 
@@ -218,17 +190,25 @@ def test_confirm_order_fail(ig, requests_mock):
     assert result == False
 
     ig_request_confirm_trade(requests_mock, fail=True)
-    result = ig.confirm_order("123456789")
-    assert result == False
+    with pytest.raises(RuntimeError) as e:
+        result = ig.confirm_order("123456789")
 
 
 def test_close_position(ig, requests_mock):
     ig_request_trade(requests_mock)
     ig_request_confirm_trade(requests_mock)
-    pos = {
-        "market": {"instrumentName": "mock"},
-        "position": {"direction": "BUY", "dealId": "123456789"},
-    }
+    pos = Position(
+        deal_id="123456789",
+        size=1,
+        create_date="mock",
+        direction=TradeDirection.BUY,
+        level=100,
+        limit=110,
+        stop=90,
+        currency="GBP",
+        epic="mock",
+        market_id=None,
+    )
     result = ig.close_position(pos)
     assert result
 
@@ -236,10 +216,18 @@ def test_close_position(ig, requests_mock):
 def test_close_position_fail(ig, requests_mock):
     ig_request_trade(requests_mock, fail=True)
     ig_request_confirm_trade(requests_mock, fail=True)
-    pos = {
-        "market": {"instrumentName": "mock"},
-        "position": {"direction": "BUY", "dealId": "123456789"},
-    }
+    pos = Position(
+        deal_id="123456789",
+        size=1,
+        create_date="mock",
+        direction=TradeDirection.BUY,
+        level=100,
+        limit=110,
+        stop=90,
+        currency="GBP",
+        epic="mock",
+        market_id=None,
+    )
     result = ig.close_position(pos)
     assert result == False
 
@@ -271,29 +259,18 @@ def test_close_all_positions_fail(ig, requests_mock):
     assert result == False
 
 
-def test_http_get(ig, requests_mock):
-    data = {"mock1": "mock", "mock2": 2}
-    requests_mock.get("http://www.mock.com", status_code=200, json=data)
-    response = ig.http_get("http://www.mock.com")
-
-    assert response is not None
-    assert isinstance(response, dict)
-    assert response == data
-
-
 def test_get_account_used_perc(ig, requests_mock):
     ig_request_account_details(requests_mock)
     perc = ig.get_account_used_perc()
 
     assert perc is not None
-    assert perc == 0
+    assert perc == 62.138354775208285
 
 
 def test_get_account_used_perc_fail(ig, requests_mock):
     ig_request_account_details(requests_mock, fail=True)
-    perc = ig.get_account_used_perc()
-
-    assert perc is None
+    with pytest.raises(RuntimeError) as e:
+        perc = ig.get_account_used_perc()
 
 
 def test_navigate_market_node_nodes(ig, requests_mock):
@@ -310,8 +287,8 @@ def test_navigate_market_node_nodes(ig, requests_mock):
     assert len(data["markets"]) == 0
 
     ig_request_navigate_market(requests_mock, fail=True)
-    data = ig.navigate_market_node("")
-    assert data is None
+    with pytest.raises(RuntimeError) as e:
+        data = ig.navigate_market_node("")
 
 
 def test_navigate_market_node_markets(ig, requests_mock):
@@ -328,42 +305,21 @@ def test_navigate_market_node_markets(ig, requests_mock):
     assert data["markets"][2]["epic"] == "KC.D.AVLN8875P.JUN.IP"
 
 
-def test_get_watchlist_list(ig, requests_mock):
-    ig_request_watchlist(requests_mock)
-    data = ig.get_watchlist("")
-    assert "watchlists" in data
-    assert len(data["watchlists"]) == 5
-    assert data["watchlists"][0]["id"] == "12345678"
-    assert data["watchlists"][1]["id"] == "Popular Markets"
-    assert data["watchlists"][2]["id"] == "Major Indices"
-    assert data["watchlists"][3]["id"] == "6817448"
-    assert data["watchlists"][4]["id"] == "Major Commodities"
-
-    ig_request_watchlist(requests_mock, fail=True)
-    data = ig.get_watchlist("")
-    assert data is None
-
-    ig_request_watchlist(requests_mock, data="mock_error.json")
-    data = ig.get_watchlist("")
-    assert data is None
-
-
 def test_get_watchlist_markets(ig, requests_mock):
+    ig_request_market_info(requests_mock)
     ig_request_watchlist(requests_mock, data="mock_watchlist_list.json")
     ig_request_watchlist(requests_mock, args="12345678", data="mock_watchlist.json")
 
     data = ig.get_markets_from_watchlist("My Watchlist")
     assert isinstance(data, list)
     assert len(data) == 3
-    assert data[0]["epic"] == "CS.D.BITCOIN.TODAY.IP"
-    assert data[1]["epic"] == "IX.D.FTSE.DAILY.IP"
-    assert data[2]["epic"] == "IX.D.DAX.DAILY.IP"
+    assert isinstance(data[0], Market)
 
     data = ig.get_markets_from_watchlist("wrong_name")
-    assert data is None
+    assert len(data) == 0
 
     ig_request_watchlist(
         requests_mock, args="12345678", data="mock_watchlist.json", fail=True
     )
     data = ig.get_markets_from_watchlist("wrong_name")
-    assert data is None
+    assert len(data) == 0